refactor: code split members page
